My Step Framework v1.9.7 1. Add atom and rss API for all of the APPS with atom.php and rss.php in the root path of each app 2. Integrate “no_log” feature with myStep::$func_ignore variable 3. Add sub-website selection feature to catalog list function of CMS app 4. Fixed a problem in captcha function with browser cache 5. Fixed a bug in SQL build function 6. Adjust request function, it will return full Predefined Variables set when no parameter is provide 7. Completely enhance bad request plugin and rename it to badReq 8. Add difference code function to manager plugin, you can download the files different with update server 9. Some other bug fix, code adjust and function optimize.

+++ lib/session/sess_mysql.class.php
@@ -12,6 +12,7 @@ class sess_mysql implements interface_session {
     }
 
     public static function open($sess_path, $sess_name) {
+        if($GLOBALS['no_log']) return true;
         $result = false;
         if(!isset(self::$setting['charset'])) self::$setting['charset'] = 'utf8';
         if(self::$cnt = mysqli_connect(self::$setting['host'], self::$setting['user'], self::$setting['password'], self::$setting['name'])) {
@@ -29,11 +30,13 @@ CREATE TABLE IF NOT EXISTS `my_session` (
     }
 
     public static function close() {
+        if($GLOBALS['no_log']) return true;
         if(rand(1,100)>95) self::gc(ini_get('session.gc_maxlifetime'));
         return true;
     }
 
     public static function read($sid) {
+        if($GLOBALS['no_log']) return '';
         if($result = mysqli_query(self::$cnt, 'SELECT value FROM my_session WHERE SID = "'.$sid.'" AND expiration > '.$_SERVER['REQUEST_TIME'])) {
             if(mysqli_num_rows($result)) {
                 $record = mysqli_fetch_assoc($result);
@@ -57,10 +60,12 @@ CREATE TABLE IF NOT EXISTS `my_session` (
     }
 
     public static function destroy($sid) {
+        if($GLOBALS['no_log']) return true;
         return mysqli_query(self::$cnt, 'DELETE FROM my_session WHERE SID='.$sid);
     }
 
     public static function gc($maxlifetime) {
+        if($GLOBALS['no_log']) return true;
         return mysqli_query(self::$cnt, 'DELETE FROM my_session WHERE expiration < ' . ($_SERVER['REQUEST_TIME'] - $maxlifetime));
     }
 }
